+++ hack/update-deepcopy.sh
@@ -8,8 +8,8 @@ CODEGEN_PKG=${CODEGEN_PKG:-$(cd ${SCRIPT_ROOT}; ls -d -1 ./vendor/k8s.io/code-ge
 verify="${VERIFY:-}"
 
 GOFLAGS="" bash ${CODEGEN_PKG}/generate-groups.sh "deepcopy" \
-  github.com/openshift/cert-manager-operator/api \
-  github.com/openshift/cert-manager-operator/api \
+  github.com/openshift/cert-manager-operator/apis \
+  github.com/openshift/cert-manager-operator/apis \
   "operator:v1alpha1" \
   --go-header-file ${SCRIPT_ROOT}/hack/empty.txt \
   ${verify}
