+++ pkg/operator/clientset/versioned/scheme/register.go
@@ -3,7 +3,7 @@
 package scheme
 
 import (
-	operatorv1alpha1 "github.com/openshift/cert-manager-operator/api/operator/v1alpha1"
+	operatorv1alpha1 "github.com/openshift/cert-manager-operator/apis/operator/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
