+++ pkg/operator/clientset/versioned/typed/operator/v1alpha1/certmanager.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"time"
 
-	v1alpha1 "github.com/openshift/cert-manager-operator/api/operator/v1alpha1"
+	v1alpha1 "github.com/openshift/cert-manager-operator/apis/operator/v1alpha1"
 	scheme "github.com/openshift/cert-manager-operator/pkg/operator/clientset/versioned/scheme"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	types "k8s.io/apimachinery/pkg/types"
