+++ pkg/operator/clientset/versioned/typed/operator/v1alpha1/fake/fake_certmanager.go
@@ -5,7 +5,7 @@ package fake
 import (
 	"context"
 
-	v1alpha1 "github.com/openshift/cert-manager-operator/api/operator/v1alpha1"
+	v1alpha1 "github.com/openshift/cert-manager-operator/apis/operator/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
