+++ pkg/operator/informers/externalversions/generic.go
@@ -5,7 +5,7 @@ package externalversions
 import (
 	"fmt"
 
-	v1alpha1 "github.com/openshift/cert-manager-operator/api/operator/v1alpha1"
+	v1alpha1 "github.com/openshift/cert-manager-operator/apis/operator/v1alpha1"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
 	cache "k8s.io/client-go/tools/cache"
 )
