+++ pkg/operator/listers/operator/v1alpha1/certmanager.go
@@ -3,7 +3,7 @@
 package v1alpha1
 
 import (
-	v1alpha1 "github.com/openshift/cert-manager-operator/api/operator/v1alpha1"
+	v1alpha1 "github.com/openshift/cert-manager-operator/apis/operator/v1alpha1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/tools/cache"
